Лекция 2.8. Создан общий базовый класс для классов помощников
Создан HelperBase, в него добавлен метод type click.
type, click внесены в NavigationHelper, SessionHelper, GroupHelper, ContactHelper.
Из класса ApplicationManager перенесен isAlertPresent в HelperBase.
