+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/ApplicationManager.java
@@ -39,15 +39,6 @@ public class ApplicationManager {
         }
     }
 
-    private boolean isAlertPresent() {
-        try {
-            wd.switchTo().alert();
-            return true;
-        } catch (NoAlertPresentException e) {
-            return false;
-        }
-    }
-
     public GroupHelper getGroupHelper() {
         return groupHelper;
     }
