+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/GroupHelper.java
@@ -4,11 +4,10 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import ru.stqa.pft.addressbook.model.GroupData;
 
-public class GroupHelper {
-    private WebDriver wd;
+public class GroupHelper extends HelperBase {
 
     public GroupHelper(WebDriver wd) {
-        this.wd = wd;
+        super(wd);
     }
 
     public void returnGroupPage() {
@@ -19,20 +18,10 @@ public class GroupHelper {
         click(By.name("submit"));
     }
 
-    private void click(By locator) {
-        wd.findElement(locator).click();
-    }
-
     public void fillGroupForm(GroupData groupData) {
         type(By.name("group_name"), groupData.getName());
         type(By.name("group_header"), groupData.getHeader());
         type(By.name("group_footer"), groupData.getFooter());
-    }
-
-    private void type(By locator, String text) {
-        click(locator);
-        wd.findElement(locator).clear();
-        wd.findElement(locator).sendKeys(text);
     }
 
     public void initGroupCreation() {
