+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/appmanager/SessionHelper.java
@@ -3,20 +3,15 @@ package ru.stqa.pft.addressbook.appmanager;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
-public class SessionHelper {
-    private WebDriver wd;
+public class SessionHelper extends HelperBase {
 
     public SessionHelper(WebDriver wd) {
-        this.wd = wd;
+        super(wd);
     }
 
     public void login(String username, String password) {
-        wd.findElement(By.name("user")).click();
-        wd.findElement(By.name("user")).clear();
-        wd.findElement(By.name("user")).sendKeys(username);
-        wd.findElement(By.name("pass")).click();
-        wd.findElement(By.name("pass")).clear();
-        wd.findElement(By.name("pass")).sendKeys(password);
-        wd.findElement(By.xpath("//input[@value='Login']")).click();
+        type(By.name("user"), username);
+        type(By.name("pass"), password);
+        click(By.xpath("//input[@value='Login']"));
     }
 }
